feat(http-client): centralize HTTP client creation with tracing into shared package
- Extract HTTP client logic with OpenTelemetry tracing into shared package
- Replace app-specific HTTP client implementations with shared createHttpClient usage
- Add @opentelemetry/api as dependency in shared package
- Simplify codebase by standardizing HTTP client creation and tracing setup
