+++ packages/shared/package.json
@@ -27,6 +27,7 @@
 	},
 	"dependencies": {
 		"@llmgateway/models": "workspace:*",
+		"@opentelemetry/api": "1.9.0",
 		"@radix-ui/react-avatar": "1.1.11",
 		"@radix-ui/react-checkbox": "^1.3.3",
 		"@radix-ui/react-collapsible": "1.1.12",
