+++ packages/shared/src/index.ts
@@ -12,6 +12,12 @@ export {
 	type HealthResponse,
 } from "./health-check.js";
 
+export {
+	createHttpClient,
+	type HttpClientOptions,
+	type HttpClientConfig,
+} from "./http-client.js";
+
 export { ModelSelector, ProviderIcons } from "./components/index.js";
 
 export { useIsMobile } from "./hooks/use-mobile.js";
